extract a CommandRouting module
